Refactored ScopeTokenToListOfRepos and moved to Github provider package
