+++ pkg/pipelineascode/match.go
@@ -3,9 +3,7 @@ package pipelineascode
 import (
 	"context"
 	"encoding/json"
-	"errors"
 	"fmt"
-	"net/url"
 	"regexp"
 	"strings"
 
@@ -13,12 +11,12 @@ import (
 	"github.com/openshift-pipelines/pipelines-as-code/pkg/apis/pipelinesascode/v1alpha1"
 	"github.com/openshift-pipelines/pipelines-as-code/pkg/matcher"
 	"github.com/openshift-pipelines/pipelines-as-code/pkg/provider"
+	"github.com/openshift-pipelines/pipelines-as-code/pkg/provider/github"
 	"github.com/openshift-pipelines/pipelines-as-code/pkg/resolve"
 	"github.com/openshift-pipelines/pipelines-as-code/pkg/secrets"
 	"github.com/openshift-pipelines/pipelines-as-code/pkg/templates"
 	tektonv1 "github.com/tektoncd/pipeline/pkg/apis/pipeline/v1"
 	"go.uber.org/zap"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
 func (p *PacRun) matchRepoPR(ctx context.Context) ([]matcher.Match, *v1alpha1.Repository, error) {
@@ -103,7 +101,7 @@ is that what you want? make sure you use -n when generating the secret, eg: echo
 	}
 
 	if p.event.InstallationID > 0 {
-		token, err := p.scopeTokenToListOfRepos(ctx, repo)
+		token, err := github.ScopeTokenToListOfRepos(ctx, p.vcx, repo, p.run, p.event, p.eventEmitter, p.logger)
 		if err != nil {
 			return nil, err
 		}
@@ -261,72 +259,4 @@ func (p *PacRun) checkNeedUpdate(tmpl string) (string, bool) {
 		return `!Update needed! you have a old basic auth secret name, you need to modify your pipelinerun and change the string "secret: pac-git-basic-auth-{{repo_owner}}-{{repo_name}}" to "secret: {{ git_auth_secret }}"`, true
 	}
 	return "", false
-}
-
-func (p *PacRun) scopeTokenToListOfRepos(ctx context.Context, repo *v1alpha1.Repository) (string, error) {
-	var (
-		listRepos bool
-		token     string
-	)
-	listURLs := map[string]string{}
-	repoListToScopeToken := []string{}
-
-	// This is a Global config to provide list of repos to scope token
-	if p.run.Info.Pac.SecretGhAppTokenScopedExtraRepos != "" {
-		for _, configValue := range strings.Split(p.run.Info.Pac.SecretGhAppTokenScopedExtraRepos, ",") {
-			configValueS := strings.TrimSpace(configValue)
-			if configValueS == "" {
-				continue
-			}
-			repoListToScopeToken = append(repoListToScopeToken, configValueS)
-		}
-		listRepos = true
-		p.logger.Infof("configured Global configuration to %v to scope Github token ", repoListToScopeToken)
-	}
-	if repo.Spec.Settings != nil && len(repo.Spec.Settings.GithubAppTokenScopeRepos) != 0 {
-		ns := repo.Namespace
-		repoListInPerticularNamespace, err := p.run.Clients.PipelineAsCode.PipelinesascodeV1alpha1().Repositories(ns).List(ctx, metav1.ListOptions{})
-		if err != nil {
-			return "", err
-		}
-		for i := range repoListInPerticularNamespace.Items {
-			splitData, err := getURLPathData(repoListInPerticularNamespace.Items[i].Spec.URL)
-			if err != nil {
-				return "", err
-			}
-			listURLs[splitData[1]+"/"+splitData[2]] = splitData[1] + "/" + splitData[2]
-		}
-		for i := range repo.Spec.Settings.GithubAppTokenScopeRepos {
-			if _, ok := listURLs[repo.Spec.Settings.GithubAppTokenScopeRepos[i]]; !ok {
-				msg := fmt.Sprintf("failed to scope Github token as repo %s does not exist in namespace %s", repo.Spec.Settings.GithubAppTokenScopeRepos[i], ns)
-				p.eventEmitter.EmitMessage(nil, zap.ErrorLevel, "RepoDoesNotExistInNamespace", msg)
-				return "", errors.New(msg)
-			}
-			repoListToScopeToken = append(repoListToScopeToken, repo.Spec.Settings.GithubAppTokenScopeRepos[i])
-		}
-		listRepos = true
-		p.logger.Infof("configured repo level configuration to %v to scope Github token ", repo.Spec.Settings.GithubAppTokenScopeRepos)
-	}
-	if listRepos {
-		repoInfoFromWhichEventCame, err := getURLPathData(repo.Spec.URL)
-		if err != nil {
-			return "", err
-		}
-		// adding the repo info from which event came so that repositoryID will be added while scoping the token
-		repoListToScopeToken = append(repoListToScopeToken, repoInfoFromWhichEventCame[1]+"/"+repoInfoFromWhichEventCame[2])
-		token, err = p.vcx.CreateToken(ctx, repoListToScopeToken, p.run, p.event)
-		if err != nil {
-			return "", fmt.Errorf("failed to scope token to repositories with error : %w", err)
-		}
-		p.logger.Infof("Github token scope extended to %v ", repoListToScopeToken)
-	}
-	return token, nil
-}
-
-func getURLPathData(urlInfo string) ([]string, error) {
-	urlData, err := url.ParseRequestURI(urlInfo)
-	if err != nil {
-		return []string{}, err
-	}
-	return strings.Split(urlData.Path, "/"), nil
 }
